+++ back-end/server.js
@@ -14,9 +14,18 @@ const PORT = 5000 || process.env.PORT;
 
 connectDB();
 
+//? User Related and Authentication Links
 app.use('/user', require('./api/routes/userRouter'));
 app.use('/auth', require('./api/routes/authRouter'));
+app.use('/visitor', require('./api/routes/visitorRouter'));
+app.use('/bldgLoc', require('./api/routes/buildingLocRouter'));
+
+//? Visitor Related Links
+app.use('/visitor', require('./api/routes/visitorRouter'));
+app.use('/visitor/companion', require('./api/routes/visitorCompRouter'));
+app.use('/visitor/logs', require('./api/routes/visitorLogsRouter'));
+app.use('/visitor/companion/logs', require('./api/routes/visitorLogsRouter'));
 
 app.listen(PORT, () => {
-    console.log("Server started on port ${PORT}");
+    console.log(`Server started on port ${PORT}`);
 })